+++ files/gitlab-cookbooks/gitlab/recipes/gitlab-shell.rb
@@ -26,9 +26,6 @@ authorized_keys = node['gitlab']['gitlab-shell']['auth_file']
 log_directory = node['gitlab']['gitlab-shell']['log_directory']
 gitlab_shell_keys_check = File.join(gitlab_shell_dir, 'bin/gitlab-keys')
 gitlab_shell_config_file = File.join(gitlab_shell_var_dir, "config.yml")
-gitlab_rails_dir = node['gitlab']['gitlab-rails']['dir']
-gitlab_rails_etc_dir = File.join(gitlab_rails_dir, "etc")
-gitlab_shell_secret_file = File.join(gitlab_rails_etc_dir, 'gitlab_shell_secret')
 
 # Creates `.ssh` directory to hold authorized_keys
 [
@@ -97,21 +94,4 @@ end
 execute "#{gitlab_shell_keys_check} check-permissions" do
   user git_user
   group git_group
-end
-
-# If SELinux is enabled, make sure that OpenSSH thinks the .ssh directory and authorized_keys file of the
-# git_user is valid.
-bash "Set proper security context on ssh files for selinux" do
-  code <<~EOS
-    semanage fcontext -a -t ssh_home_t '#{ssh_dir}(/.*)?'
-    semanage fcontext -a -t ssh_home_t '#{authorized_keys}'
-    semanage fcontext -a -t ssh_home_t '#{gitlab_shell_config_file}'
-    semanage fcontext -a -t ssh_home_t '#{gitlab_shell_secret_file}'
-    restorecon -R -v '#{ssh_dir}'
-    restorecon -v '#{authorized_keys}' '#{gitlab_shell_config_file}'
-    # On new installs, the gitlab_shell_secret file may not exist until the
-    # gitlab-rails recipe runs, so we can safely move along if the file doesn't exist.
-    restorecon -v -i '#{gitlab_shell_secret_file}'
-  EOS
-  only_if "id -Z"
 end