+++ CHANGELOG.md
@@ -9,6 +9,8 @@ omnibus-gitlab repository.
 - Disable SSL compression by default when using gitlab-psql or gitlab-geo-psql
 - Increase Sidekiq RSS memory limit from 1 GB to 2 GB
 - Reload sysctl if a new symlink is created in /etc/sysctl.d
+- Update libiconv to 1.15
+- Update libiconv zlib and makedepend build recipe to improve build cache reuse
 
 11.2.1